Added workaround empty service catalog build.
The workaround is to get an auth token manually using urllib2.
The tenant id is then added to the api url. Those two pieces
of information (token and api url with tenant) are used to create
driver instances.

Tenants with this bug will need to add the datasource tenant
to the configuration.